Add support for inspecting only specified packages.
Currently, the plugin inspects all the data classes in a project.
It could be that you only want to inspect certain data classes.

This commit provides this functionality. Now, you can specify the packages which needs to be inspected via the settings.

If no packages are provided then the plugin will continue to inspect all the data classes.